multiModelImageAgent/requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & LangGraph
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.0
langgraph==0.0.20

# HTTP 客户端
httpx==0.25.2
aiohttp==3.9.1

# 数据库 (SQLite)
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.0

# 缓存 (可选)
redis==5.0.1

# 工具
python-dotenv==1.0.0
loguru==0.7.2
cryptography==41.0.7
PyJWT==2.12.1

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
